requirements.txt
numpy==1.21.4
pandas==1.3.4
PyYAML==6.0
requests==2.26.0
XlsxWriter==3.0.2
